db/requirements.txt
django
djangorestframework
django-cors-headers
django_countries
gunicorn
psycopg2-binary
pylint
pylint-django
